requirements.txt
annotated-types==0.7.0
anyio==4.4.0
asgiref==3.8.1
Brotli==1.1.0
certifi==2024.7.4
coloredlogs==15.0.1
distro==1.9.0
dj-database-url==2.2.0
Django==5.0.2
django-cors-headers==4.4.0
django-crispy-forms==1.14.0
django-environ==0.11.2
exceptiongroup==1.2.2
flatbuffers==24.3.25
groq==0.9.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
humanfriendly==10.0
idna==3.7
mpmath==1.3.0
numpy==1.26.4
onnx==1.15.0
onnxruntime==1.18.0
opencv-python==4.10.0.84
packaging==24.1
pillow==10.4.0
protobuf==5.27.2
psycopg2==2.9.9
pydantic==2.8.2
pydantic_core==2.20.1
pyreadline3==3.4.1
python-dotenv==1.0.1
sniffio==1.3.1
sqlparse==0.5.0
sympy==1.12.1
typing_extensions==4.12.2
tzdata==2024.1
whitenoise==6.7.0
